Supply ancestors (array) to the validators.
Previously, it supplied the root level data (`dataRoot`) to the
validators along with the field data to validate. However, it was
limiting if the data was deeply nested.

Now, it supplies the field data as well as all its ancestors to the
validators. So you can validate the field according to its parent's data
or its grand parent's data, and so on.