`isRequired()` checks for `undefined` and `null`.
Previously, `isRequired()` checked if the value provided is `undefined`
or not (see 78ff21). So providing `null` would have passed
`isRequired()` check.

However, I decided to change it to include `null` check as well
because of the following reason:

```js
JSON.stringify([undefined, 2]) === "[null,2]"
```

This is problematic when you do server-side validation. Even if client
sends `undefined` in an array, it will be converted to `null` and pass
the `isRequired()` check (hence, will be validated against the provided
field validators).

So with this commit, `null` value won't be validated against the
provided field validators.